+++ src/Model/Entity/Page.php
@@ -23,11 +23,11 @@
 namespace MeCms\Model\Entity;
 
 use Cake\ORM\Entity;
-use MeCms\Model\Entity\Traits\PreviewAccessorTrait;
 
 /**
  * Page entity
  * @property int $id
+ * @property int $category_id
  * @property string $title
  * @property string $subtitle
  * @property string $slug
@@ -36,11 +36,10 @@ use MeCms\Model\Entity\Traits\PreviewAccessorTrait;
  * @property bool $active
  * @property \Cake\I18n\Time $created
  * @property \Cake\I18n\Time $modified
+ * @property \MeCms\Model\Entity\PagesCategory $category
  */
 class Page extends Entity
 {
-    use PreviewAccessorTrait;
-
     /**
      * Fields that can be mass assigned using newEntity() or patchEntity()
      * @var array
@@ -48,12 +47,7 @@ class Page extends Entity
     protected $_accessible = [
         '*' => true,
         'id' => false,
+        'preview' => false,
         'modified' => false,
     ];
-
-    /**
-     * Virtual fields that should be exposed
-     * @var array
-     */
-    protected $_virtual = ['preview'];
 }